+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Improvements:
  * reset aarch64 feature string as +all;
  * remove unused assembler llvm-mc logic to improve performance;
+ * add new binary from memory object apis;
 
 ## v0.1.0
 Initial release.